fix: continue NASA Power of Ten safety hardening — assertions, function decomposition, and bounds checks
Adds Q_ASSERT pre/post-conditions across API, DataModel, IO, CSV, Dashboard, and export subsystems.
Refactors API::Server::onDataReceived into focused helper methods (validateRateLimits, validateJsonMessage,
handleJsonMessage, processRawJsonCommand, processBufferedJson, processJsonLine, processRawLine) to meet
the 60-line function limit. Replaces unsafe pointer-to-int session IDs with a monotonic atomic counter.
Adds iteration guards on unbounded loops and improves error handling throughout.

+++ app/src/DataModel/ProjectModel.cpp
@@ -2342,7 +2342,7 @@ void DataModel::ProjectModel::addGroup(const QString& title, const SerialStudio:
 bool DataModel::ProjectModel::setGroupWidget(const int group,
                                              const SerialStudio::GroupWidget widget)
 {
-  if (group < 0 || group >= m_groups.size()) [[unlikely]]
+  if (group < 0 || group >= static_cast<int>(m_groups.size())) [[unlikely]]
     return false;
 
   auto& grp          = m_groups[group];
